+++ rs/pocket_ic_server/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- New GET endpoint `/http_gateway` listing all HTTP gateways and their details.
 
 
 ## 5.0.0 - 2024-07-22